Nurit PR: Changes to harness funcs.

+++ certora/harness/IATokenWithPool.sol
@@ -9,7 +9,7 @@ interface IATokenWithPool {
 
     function balanceOf(address account) external view returns (uint256);
 
-    function transfer(address recipient, uint128 amount)
+    function transfer(address recipient, uint256 amount)
         external
         returns (bool);
